Add MessageTypes.js
